Arreglando back

+++ src/main/java/com/portfolio/lngport/Repository/IPersonaRepository.java
@@ -8,7 +8,7 @@ import org.springframework.stereotype.Repository;
 
 
 @Repository
-public interface IPersonaRepository extends JpaRepository <Persona,Integer>  {
+public interface IPersonaRepository extends JpaRepository <Persona,Long>  {
     public Optional<Persona> findByNombre(String nombre);
     public boolean existsByNombre(String nombre);
